Rename lazy_loader.js to ensure_loaded.js. Also clean up the CanHaveAPIData store mixin, meant to do these separetly but forgot to commit, and there's to many files to split the commit now, oops
