+++ client/src/TagExplorer/TagExplorer.js
@@ -1,7 +1,7 @@
 import "src/explorer_common/explorer-styles.scss";
 import { SpinnerWrapper } from "src/components/index.js";
 
-import { ensure_loaded } from "src/core/lazy_loader.js";
+import { ensure_loaded } from "src/core/ensure_loaded.js";
 import { StandardRouteContainer } from "src/core/NavComponents.js";
 
 import {
