+++ client/src/panels/panel_declarations/results/result_drilldown/results_scheme.js
@@ -1,7 +1,6 @@
 import { createSelector } from "reselect";
 
-import { ensure_loaded } from "src/core/lazy_loader.js";
-
+import { ensure_loaded } from "src/core/ensure_loaded.js";
 
 import { AbstractExplorerScheme } from "src/explorer_common/abstract_explorer_scheme";
 import { filter_hierarchy } from "src/explorer_common/hierarchy_tools.js";
